requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
openai>=2.14.0
python-dotenv>=1.0.1
pypdf>=5.3.1
rlms>=0.1.1
